requirements.txt
dash-bootstrap-components==1.4.2
PyYAML==6.0.1
openai==0.27.8
langchain[llms]==0.0.266
dash==2.12.1
tiktoken==0.4.0
python-dotenv==1.0.0
pydantic==1.10.12
llama-index==0.8.5.post1
redis==5.0.0
nltk==3.8.1
psycopg==3.1.10
SQLAlchemy==2.0.20
psycopg2==2.9.7
